gpu: cross-platform surfaces with GLFW;
Originally I wanted to create the surfaces manually using native handles,
but it seems to be impossible to create a metal surface using only the
GLFW-provided handles.  Instead you are forced to go through GLFW's
surface create function.  So for now we are going to switch to passing
in callbacks to core/gpu for requesting instance extensions and creating
the surface.